Create test_melt.py
